fix: Fabric crash (nested transaction), FluidTank refactor, Forge ModelData
- Fix Fabric crash when pipe connects to ChemPlant: ChemPlantFabricFluidStorage
  now uses tank.getStorage().insert/extract directly with the parent transaction
  instead of addCloseCallback + fillMb (which opened a nested outer transaction)
- Separate FluidLoaderStandard and FluidLoaderInfinite into their own files
  (matching 1.7.10 structure) — clean FluidTank.java without nested loader classes
- Fix Forge ModelData: add requestModelDataUpdate() in load() for client-side
  so the fluid tank texture updates when synced from server
